Fix parallel owner selection in STK
The comment states that the owner of an entity is lowest rank process that shares
that entity.  The code, however, was actually picking the highest rank process.
This fixes the behavior to match the comment (which is also the behavior of framework.)

Fixup test for fix to entity processor owner.
An entity is no longer present on a certain proc due to ownership changes.

(From Sierra)