MueLu: fixing damping factor format in XML decks
There were a number of XML decks specifying the damping factor like so:

     <Parameter name="Damping factor"                      type="double" value="4./3"/>

The XML parser interprets this as a damping factor of 4!  We should expect the parser
to understand mathematical expressions, I guess.

This commit changes all such occurences to 1.333.

Thanks to Dr. PTLin for helping to locate this.