Ifpack2: SparseContainer: Attempt to address gcc 4.1.2 build errors.
The build errors might have something to do with the compiler not
being able to deduce that the result of *(RCP<MV>) can be cast to a
const MV&.  The errors only occur when calling Inverse_->apply, on
the first argument of that call.  I added explicit casts of the
result of *(RCP<MV>) to const MV&, for all places where the code was
calling Inverse_->apply.  Hopefully that will fix the issue.

Build/Test Cases Summary
Enabled Packages: Ifpack2
0) MPI_DEBUG => passed: passed=22,notpassed=0 (0.48 min)
1) SERIAL_RELEASE => passed: passed=18,notpassed=0 (0.05 min)