Ifpack2: SparseContainer: Fixed diagonal scaling in weightedApply.
weightedApply was doing the diagonal scaling (elementWiseMultiply) with
the whole diagonal Vector D, instead of the permuted subset that it
should be using.  The "permuted subset" refers to the domain and range
of the InverseType operator.  This commit fixes that.  It makes the code
correct, and might also improve performance for small diagonal blocks.
Of course, the "small diagonal blocks" case really needs DenseContainer
for best performance, which Ifpack2 doesn't have yet.

Build/Test Cases Summary
Enabled Packages: Ifpack2
0) MPI_DEBUG => passed: passed=22,notpassed=0 (0.43 min)
1) SERIAL_RELEASE => passed: passed=18,notpassed=0 (0.05 min)