MueLu: fixed DirectSolver test
The test tries to construct DirectSolver with different types and
parameter lists. The problem is that the test never worked, as the
constructors of the DirectSolver and Amesos[2]Smoother were empty, so
the arguments were never checked. Now that I moved most of the staff
into the constructors, the test suddenly broke.

I fixed the test. It is much more complex now, as the behaviour of the
DirectSolver constructor depends on the compilation (as if we compiled
MueLu with Amesos2 support, and whether Amesos2 actually has SuperLu).