Changed logic to not enable package for no-implicit enable repo even if tests are enabled (VRI Kanban #2920)
With this change, the TribitsCTestDriverCore.cmake code should no longer
enable for testing a modified package that is from a repo that has its
packages by default to not be implicitly enabled.  This is needed in large
projects like CASL VERA that has a lot of packages that don't need to be
tested as part of VERA but yet are changing during the testing day as the CI
server is running.  We really only want to test the impact of downstream
packages for modified packages.

I also changed some names of some macros as part of this.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Disabled Packages: CTrilinos,ForTrilinos,PyTrilinos,Didasko,Mesquite,Phdmesh,Pliris,Claps
Enabled all Packages
0) MPI_DEBUG => passed: passed=1456,notpassed=0 (49.37 min)
1) SERIAL_RELEASE => passed: passed=1432,notpassed=0 (28.34 min)
Other local commits for this build/test group: 2a0d267