Kokkos: Fixed build issue with gcc 4.2.1 version of std::pair.
gcc 4.2.1 is the version of gcc used by default on Mac OS X, so this
could be considered a Mac fix.  Apparently, with that version of the C++
Standard Library, "std::pair<unsigned, unsigned> (0u, 0u)" as a default
argument of a constructor doesn't work, but "std::make_pair (0u, 0u)"
does work.