MueLu: fix nightly testing of epetra/tpetra only
This commit explicitly disables packages by passing options of the form "-DTrilinos_ENABLE_Foo:BOOL=OFF" to the cmake line.

Previously, the scripts to test MueLu with only Epetra or with only Tpetra were incorrect, and not doing what was intended.
They were using the variable EXTRA_EXCLUDE_PACKAGES.  However, this variable instructs the testing system to not enable package
FOO *directly*, but FOO can still be enabled through an optional dependency on an enable package.

Thanks to Brent P. for setting me straight on this.