Cleaning up TRIBITS_PROJECT_IMPL() macro, make packaging optional.
I have created slave macros for the big chunks of code in this main
TRIBITS_PROEJCT_IMPL() function.  This is so that you can see the global logic
for when things occur.  This will be used to define the algorihtm for when
things happen.

There is a non-trivial amount of work involved in setup up for a distribution
with CPack.  Since most developers will not be creating a source distriubtion,
it makes no sense for them to have to pay this cost all the time when they
configure.