SEACAS: conjoin - Added interpart_minimum_time_delta {delta} option
Specifies the minimum time delta between the last time on one database
and the first time on the next database for those times to be considered
as different times and transferred to the output database.

The default value is 0.0 which means that all non-equal times will
be output.

Addresses desire expressed in trac #7499

(From Sierra)