Stokhos:  Performance improvements for SG SpMv for MIC architecture.
I added vector instructions for the MIC architecture, added versions of
the CrsProductTensor algorithm that use auto and manual vectorization,
and created a version of the GPU algorithm for the MIC that processes
multiple FEM columns at a time to reduce global reads of the Cijk
tensor.  Together the vectorization + GPU algorithm improve performance
on the MIC considerably (with the same cache-size drawbacks as on the
GPU).  For the CPU, the GPU algorithm doesn't improve the performance,
but turning on vectorization on SandyBridge does help a little.