Kokkos: LinAlg: Added CMake-enabled unit test (CrsMatrixUnitTest).
Currently, the test only verifies that CrsMatrix compiles and that it
doesn't throw an exception on construction.  I just wanted something
that would run automatically with CMake.  The current tests are much
more extensive, but are not plugged into CMake, so they aren't run
automatically by the check-in test script.  We will eventually want to
plug at least the existing unit tests (if not the performance tests)
into CMake, but this should suffice for now.