Intrepid: Fix boost dependency
MiniTensor has a dependency on boost that wasn't being specified. It was working
because Teuchos has an optional dependency on boost which was pulling in the
necessary dependencies for Intrepid. This makes the dependency on boost explicit
to ensure the necessary setup is done for Intrepid.

Build/Test Cases Summary
Enabled Packages: Intrepid
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=131,notpassed=0 (18.08 min)
1) SERIAL_RELEASE => passed: passed=131,notpassed=0 (16.95 min)