MueLu: handle direct solver type "Umfpack" within Amesos2Solver
Umfpack is not available with Amesos2. But due to the recent changes in the solver/smoother
framework we check whether we have to build an Epetra or Tpetra based solver/smoother.
However if we request a Umfpack direct solver (for Epetra) we have to make sure that we
do not run into an exception within Amesos2Solver.

I'm not really happy with that solution. This issue needs some more discussion with the MueLu devs.