MueLu: fixing tests for the new default max coarse size
Many tests did not specify max coarse size, so when I increased the
limit, they suddenly did not construct enough levels for tests. Now,
tests explicitly specify that. There are probably some more broken
tests, and this should be specified for all tests, but I'm not going to
hunt for them now.