Panzer: Enabling turning on/off of subcell construction for CubeHex mesh
This routine in stk is poorly written, and has been causing
memory issues. Hopefully this is the main issue with regard
to mesh construction and will allow better extreme-scale simulations
using just this mesh. If other meshes are desired then their
factories will need to be modified.

To turn on/off add one parameter list to the mesh sublist

        <ParameterList name="Inline Mesh">
            <Parameter name="Mesh Dimension" type="int" value="3" />
            <ParameterList name="Mesh Factory Parameter List">
                <Parameter name="X Blocks" type="int" value="1" />
                <!--
                   ... More parameters here
                 -->

                <!-- Turn on/off (true/false) subcell construction -->
                <Parameter name="Build Subcells" type="bool" value="false" />
            </ParameterList>
        </ParameterList>