Tpetra: CrsGraph, CrsMatrix: Fixed Bug 5978.
Thanks to Clark Dohrmann for discovering this bug and providing a test
case.  Clark's test uses 2 MPI processes.  Process 0 gets a nonzero
number of rows in the row Map, and Process 1 gets zero rows in the row
Map.  The test then creates a graph with that row Map, and with a column
Map equal to that row Map.  Then, only Process 0 calls
insertLocalIndices() to insert a nonzero number of entries into the
graph.  Finally, the test calls fillComplete, supplying domain and range
Maps equal to the row Map.

On both processes, the graph starts neither locally nor globally
indexed.  In the test, only Process 0 calls insertLocalIndices(), since
Process 1 has zero rows.  The very first call to insertLocalIndices() on
Process 0 calls allocateValues(LocalIndices, GraphNotYetAllocated).
This in turn calls allocateIndices(LocalIndices).  This marks the graph
as locally indexed, and as _not_ globally indexed:

indicesAreLocal_  = (lg == LocalIndices);
indicesAreGlobal_ = (lg == GlobalIndices);

When the test calls fillComplete() on the matrix, indices have only been
allocated (i.e., allocateIndices() has been called) only on Process 0,
not on Process 1.  This confuses later parts of fillComplete that expect
the "allocated" and "locally vs. globally indexed" states to be
consistent on all processes.  The result: MPI either hangs in a
collective, or the collective raises an MPI error (e.g.,
MPI_ERR_TRUNCATE).

Here is the fix implemented in this commit:

1.  Change the call to allocateValues in fillComplete (both in CrsGraph
and in CrsMatrix), so that its first argument is LocalIndices if the
matrix has a column Map, or GlobalIndices if the matrix does not have a
column Map.  That way, indices are always allocated on all processes
after this point.

2. That change alone causes processes with an empty graph
(insertLocal{Values,Indices} was never called) to have a different
locally vs. globally indexed state than processes with a nonempty graph.
I fixed this by making makeIndicesLocal an entirely noncollective
method.  In particular, it no longer calls computeIndexState, which does
an all-reduce to line up the locally vs. globally indexed states on all
processes.

The all-reduce mentioned in #2 was entirely unnecessary, because at the
end of fillComplete, the graph / matrix is always locally indexed and
not globally indexed.  Thus, this fix has the pleasant side effect of
removing an all-reduce from fillComplete.