TrilinosCouplings:  Set reasonable workset size in matrix assembly.
The workset size was being set to the whole problem size, which was way
too big for a large problem.  I think it was set this way because there
was a bug in the graph generation when the workset size didn't evenly
divide the problem size.  This has now been fixed.