TriKota:  Check each #define is non-empty before setting.
There is some cmake code in TriKota to extract all of Dakota's #define's
and set them in a Trilinos CMake variable and CPP macro.  With CMake
2.8.12, I was getting one of the defines as being empty, leaving a blank
define in TriKota_config.h, and causing a compiler error.  I fixed this
by checking whether each define is non-empty before adding it to the
list.