For problems with element death and contact a huge amount of time was being spent in
stk::mesh::copy_entity_fields_callback

Pull out all the stops to optimize this routine as much as possible.  However, this only improves the performance
of the routine by about 10% as the majority of the call is the memcpy.

The larger problem is there is far too much object movement generated by the skinning algorithm.  Next Will try
to refactor at the algorithm level to framework roster update calls.

(From Sierra)