Tpetra: Added the second of several "lessons."
See my previous commit for a description of the lessons.  This commit
adds the second of those lessons, the "Map and Vector" example.  It
explains how Map works, how to create Vectors, and how to read and
modify the entries of a Vector.

This commit also adds a README to tpetra/example/.  The README explains
how to add a new Lesson.

Build/Test Cases Summary
Enabled Packages: Tpetra
0) MPI_DEBUG => passed: passed=64,notpassed=0 (1.29 min)
1) SERIAL_RELEASE => passed: passed=49,notpassed=0 (0.07 min)
Other local commits for this build/test group: cb85132, 97175cc, 7b2b381