Ifpack2: Deprecated "overlap as int" version of AdditiveSchwarz ctor.
It's silly for AdditiveSchwarz's constructor to take the overlap level
as an integer, since its ParameterList may already include the "schwarz:
overlap level" parameter.  Thus, I've deprecated the two-argument
version of the constructor (that takes the overlap level as the second
argument).  I've also done the same for the three-argument version of
the analogous Factory::create() method, and fixed all use of deprecated
methods in Ifpack2.