Ifpack2: Fixed IdentitySolver.
IdentitySolver didn't quite have the right semantics with respect to
setMatrix(), initialize(), and compute().  I also took the opportunity
to document it and to extend its behavior so it makes sense even if the
domain and range of the input matrix are not the same.

Build/Test Cases Summary
Enabled Packages: Ifpack2
0) MPI_DEBUG => passed: passed=56,notpassed=0 (0.50 min)
1) SERIAL_RELEASE => passed: passed=52,notpassed=0 (0.07 min)