Kokkos: Improved texture object implementation for CudaView
Using the size of the scalar type, it was possible to have only two
implementations of the CudaTextureFetch class. One if Scalar type can't
be supported with TextureObjects, and one if it is. The new implementation
can use TextureObjects for any scalar type with size of 4,8 or 16 bytes.
Essentially the workaround for doubles with reinterpret_cast is now used
always.