Stokhos:  Make Cuda unit tests run in serial.
Suddenly my K20X has issues running multiple Cuda tests simultaneously,
so now the Cuda tests are set to run in serial, ie, one at a time in
CTest.  This is the right thing to do though because Cuda cards can
be configured for exclusive access, and our tests should work in that
case.