MueLu: Graph, LWGraph: speed up isLocalNeighborVertex by an order of magnitude
Now we cache in the min and max local indices so that we don't have to
call corresponding map->isLocalNodeIndex(). It seems that that function
(or some on on the way Tpetra->Xpetra->MueLu) is not getting inlined
which leads to terrible performance in all aggregation routines.

Introducing this caching sped up one UncoupledAggregationAlgorithm from
0.70 to 0.04 for one of the problems.

It is also the case that Epetra's version of the routine is much slower
than Tpetra's, but this caching makes us independent of that.