Ifpack2: IlukGraph: Fixed potential segfaults.
constructFilledGraph() was doing the usual &x[0] thing for
Teuchos::Array x, in a few places.  This is a bad idea if x.size() == 0.
I fixed the issue and simplified the code at the same time, yay!  Of
course this could only come up if the graph didn't have entries in the
upper or lower triangle for that row, but I still prefer correct code
that probably never executes to incorrect code that probably never
executes.

I also fixed some issues involving IlukGraph's GraphType template
parameter.  As the class was written before this commit, GraphType could
only be Tpetra::CrsGraph, because the class was creating internal graphs
and assuming Tpetra::CrsGraph's constructor and methods.  In this
commit, I changed all of the graphs that IlukGraph creates to be
Tpetra::CrsGraph explicitly.  I'll have to go over the class'
expectations of the input graph more carefully in order to check whether
I can relax it to a Tpetra::RowGraph, or whether I must leave it as a
Tpetra::CrsGraph.  If I can relax the input graph to a Tpetra::RowGraph,
then I can force the class to require that GraphType be a
Tpetra::RowGraph specialization, which is the most general.

All of this is meant to be preparation for making RILUK implement
Details::CanChangeMatrix.  It's going to be tricky.