Tpetra tests: Make them work with the CudaWrapperNode
when using the CudaWrapperNode everything has to be compiled with NVCC. To make
cmake do that a .cu file has to be added for each .cpp file. It has the same name
but ends on *_CU.cu. The _CU is there to get different object file names for the cpp
variant and the cu variant. The .cu file only includes the .cpp file. Also some ifdefs
are added so that only one of the two variants gets compiled. When KokkosCompat is activated
and Cuda is activated the .cu file is actually compiled (the .cpp file is in this case empty due to
define guards) otherwise the .cpp file.