Tpetra:  Port DistObjectKA to KokkosRefactor for MultiVector.
DistObjectKA is the Kokkos "Array" version of DistObject that expects
each client object to permute/pack/unpack directly on the device using
Kokkos::View.  I ported this implementation to the KokkosRefactor
MultiVector by converting it to a partial specialization of DistObject
for the KokkosCompat wrapper node.  Since not all Tpetra classes
implement the new view-based API, the old Teuchos array-based interface
is still there as well, with a virtual member function to determine
which interface to use.  Once all Tpetra objects derived from DistObject
are converted to Kokkos View's, the old one can go away.

I also took the opportunity to rewrite the MultiVector
permute/pack/unpack kernels to work with the 2-D view directly, instead
of reinterpreting it as a 1-D view.  This makes the kernels simpler.

For the time being, the old DistObjectKA is still there, since I and a
few other people use it.  It should go away soon though.