Belos: GCRODRSolMgr now builds with complex ScalarType.
It does NOT run with complex ScalarType.  GCRODRSolMgr's constructor
throws if ScalarType is complex.  The following new test exercises this:

packages/belos/test/RealSolverManager/test_gcrodr_complex.cpp

The point is that we want SolverFactory to compile for both real and
complex ScalarType.  This is not currently true, due to solvers like
GCRODRSolMgr, whose implementations do not compile for complex
ScalarType.  This commit adds partial specializations for GCRODRSolMgr
that ensures it compiles for both real and complex ScalarType.  The
complex ScalarType partial specialization is trivial, and its
constructor throws std::logic_error.  The new class
Details::RealSolverManager makes it easy to write such partial
specializations, and GCRODRSolMgr is an example of how to do this.