Adding new test to ShyLU
This commit adds a new test to the ShyLU package, checking the
functionality related to IQR.

Due to the dependence on ShyLU, IQR and the Ifpack_DynamicFactory class,
the test needs the following CMake variables enabled:
Ifpack_ENABLE_DYNAMIC_FACTORY:BOOL=ON
Ifpack_ENABLE_PARALLEL_SUBDOMAIN_SOLVERS:BOOL=ON

The test does NOT fail if these variables are not declared, however.

The commit also patches the existing test of ShyLU to function also
when the variables are enabled.