Ifpack2: RILUK: Added checks to prep for implementing CanChangeMatrix.
I want RILUK to implement Ifpack2::Details::CanChangeMatrix.  This
will make it a valid subdomain solver for AdditiveSchwarz.  In
preparation for this, I have added checks to AdditiveSchwarz to ensure
that users are scolded (with an informative error message) for calling
methods that require the input matrix to be nonnull, if the input
matrix is null.